@@ -1,33 +1,23 @@
 {
-  "name": "vue",
-  "version": "1.0.0",
-  "description": "",
-  "main": "webpack.config.dev.js",
+  "name": "vite-vue-starter",
+  "version": "0.0.0",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "webpack-dev-server --progress --config webpack.config.dev.js --host 192.168.15.65 --open --port 6008 --hot"
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
   },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "mint-ui": "^2.2.12",
-    "moment": "^2.19.3",
-    "vue": "^2.5.9",
-    "vue-preview": "^1.0.5",
-    "vue-resource": "^1.3.4",
-    "vue-router": "^3.0.1",
-    "vuex": "^3.0.1"
+    "axios": "^0.21.1",
+    "element-plus": "^1.0.2-beta.44",
+    "normalize.css": "7.0.0",
+    "vue": "^3.0.5",
+    "vue-router": "^4.0.4",
+    "vuex": "^4.0.0"
   },
   "devDependencies": {
-    "css-loader": "^0.28.7",
-    "file-loader": "^1.1.5",
-    "html-webpack-plugin": "^2.30.1",
-    "style-loader": "^0.19.0",
-    "url-loader": "^0.6.2",
-    "vue-loader": "^13.5.0",
-    "vue-template-compiler": "^2.5.9",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.5"
+    "@vitejs/plugin-vue": "^1.2.2",
+    "@vue/compiler-sfc": "^3.0.5",
+    "vite": "^2.2.3",
+    "vite-plugin-style-import": "^0.10.0"
   }
 }
